// src/services/types/serverTypes.ts
export type IntegrationEvent<T> = {
    type: string;
    body: T;
};

export enum CoinEvent {
    INVALID_EVENT = "INVALID_EVENT",
    READ_COIN_SUCCESS = "READ_COIN_SUCCESS",
    READ_COIN_FAILURE = "READ_COIN_FAILURE",
    GET_COIN_SUCCESS = "GET_COIN_SUCCESS",
    GET_COIN_FAILURE = "GET_COIN_FAILURE",
    ADD_COIN_SUCCESS = "ADD_COIN_SUCCESS",
    ADD_COIN_FAILURE = "ADD_COIN_FAILURE",
    CREATE_COIN_SUCCESS = "CREATE_COIN_SUCCESS",
    CREATE_COIN_FAILURE = "CREATE_COIN_FAILURE",
    UPDATE_COIN_SUCCESS = "UPDATE_COIN_SUCCESS",
    UPDATE_COIN_FAILURE = "UPDATE_COIN_FAILURE",
}

export enum ProductEvent {
    INVALID_EVENT = "INVALID_EVENT",
    BUY_PRODUCT_SUCCESS = "BUY_PRODUCT_SUCCESS",
    BUY_PRODUCT_FAILURE = "BUY_PRODUCT_FAILURE",
    GET_PRODUCT_SUCCESS = "GET_PRODUCT_SUCCESS",
    GET_PRODUCT_FAILURE = "GET_PRODUCT_FAILURE",
    SET_PRODUCT_QUANTITY_SUCCESS = "SET_PRODUCT_QUANTITY_SUCCESS",
    SET_PRODUCT_QUANTITY_FAILURE = "SET_PRODUCT_QUANTITY_FAILURE",
    SET_PRICE_SUCCESS = "SET_PRICE_SUCCESS",
    SET_PRICE_FAILURE = "SET_PRICE_FAILURE",
    CREATE_PRODUCT_SUCCESS = "CREATE_PRODUCT_SUCCESS",
    CREATE_PRODUCT_FAILURE = "CREATE_PRODUCT_FAILURE",
    ADD_PRODUCT_SUCCESS = "ADD_PRODUCT_SUCCESS",
    ADD_PRODUCT_FAILURE = "ADD_PRODUCT_FAILURE",
    UPDATE_PRODUCT_QTY_SUCCESS = "UPDATE_PRODUCT_SUCCESS",
    UPDATE_PRODUCT_QTY_FAILURE = "UPDATE_PRODUCT_FAILURE",
}

export default interface IMaintainer {
    _id: string;
    username: string;
    password: string;
}